# Frontend/backend_openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "Valio Aimo Customer Portal API",
      "description": "Backend API for Zero-Fail Logistics Customer Portal",
      "version": "2.0.0"
    },
    "paths": {
...
      "/customer/orders/": {
        "get": {
          "tags": [
            "customer-orders"
          ],
          "summary": "Get Orders",
          "description": "Get all customer orders with product information",
          "operationId": "get_orders_customer_orders__get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/OrderWithDetails"
                    },
                    "type": "array",
                    "title": "Response Get Orders Customer Orders  Get"
                  }
                }
              }
            }
          },
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ]
        },
        "post": {
          "tags": [
            "customer-orders"
          ],
          "summary": "Place Order",
          "description": "Place order from cart",
          "operationId": "place_order_customer_orders__post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPlaceFromCart"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          },
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ]
        }
      },
      "/customer/orders/{order_id}": {
        "get": {
          "tags": [
            "customer-orders"
          ],
          "summary": "Get Order",
          "description": "Get specific order details with product information",
          "operationId": "get_order_customer_orders__order_id__get",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "order_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Order Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/OrderWithDetails"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/customer/orders/{order_id}/messages": {
        "get": {
          "tags": [
            "customer-orders"
          ],
          "summary": "Get Order Messages",
          "description": "Get messages for an order",
          "operationId": "get_order_messages_customer_orders__order_id__messages_get",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "order_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Order Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Message"
                    },
                    "title": "Response Get Order Messages Customer Orders  Order Id  Messages Get"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "customer-orders"
          ],
          "summary": "Send Order Message",
          "description": "Send message for an order",
          "operationId": "send_order_message_customer_orders__order_id__messages_post",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "order_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Order Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageCreate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/admin/orders/": {
        "get": {
          "tags": [
            "admin-orders"
          ],
          "summary": "Get All Orders",
          "description": "Get all orders with product information",
          "operationId": "get_all_orders_admin_orders__get",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "skip",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "default": 0,
                "title": "Skip"
              }
            },
            {
              "name": "limit",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "default": 100,
                "title": "Limit"
              }
            },
            {
              "name": "status",
              "in": "query",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Status"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/OrderWithDetails"
                    },
                    "title": "Response Get All Orders Admin Orders  Get"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/admin/orders/{order_id}": {
        "get": {
          "tags": [
            "admin-orders"
          ],
          "summary": "Get Order",
          "description": "Get specific order details with product information",
          "operationId": "get_order_admin_orders__order_id__get",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "order_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Order Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/OrderWithDetails"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/admin/orders/{order_id}/status": {
        "put": {
          "tags": [
            "admin-orders"
          ],
          "summary": "Update Order Status",
          "description": "Update order status with automatic alerts and invoice management",
          "operationId": "update_order_status_admin_orders__order_id__status_put",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "order_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Order Id"
              }
            },
            {
              "name": "status",
              "in": "query",
              "required": true,
              "schema": {
                "$ref": "#/components/schemas/OrderStatus"
              }
            },
            {
              "name": "notes",
              "in": "query",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Notes"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/admin/orders/{order_id}/replace-product": {
        "post": {
          "tags": [
            "admin-orders"
          ],
          "summary": "Replace Product With Substitute",
          "description": "Replace a product in order with customer's selected substitute",
          "operationId": "replace_product_with_substitute_admin_orders__order_id__replace_product_post",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "order_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Order Id"
              }
            },
            {
              "name": "line_id",
              "in": "query",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Line Id"
              }
            },
            {
              "name": "substitute_product_code",
              "in": "query",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Substitute Product Code"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/OrderChange"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/admin/orders/{order_id}/messages": {
        "get": {
          "tags": [
            "admin-orders"
          ],
          "summary": "Get Order Messages",
          "description": "Get messages for an order",
          "operationId": "get_order_messages_admin_orders__order_id__messages_get",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "order_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Order Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Message"
                    },
                    "title": "Response Get Order Messages Admin Orders  Order Id  Messages Get"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "admin-orders"
          ],
          "summary": "Send Order Message",
          "description": "Send message for an order",
          "operationId": "send_order_message_admin_orders__order_id__messages_post",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "order_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Order Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageCreate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
...
        "Message": {
          "properties": {
            "message_id": {
              "type": "integer",
              "title": "Message Id"
            },
            "order_id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order Id"
            },
            "claim_id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Claim Id"
            },
            "sender_type": {
              "$ref": "#/components/schemas/MessageSenderType"
            },
            "sender_id": {
              "type": "string",
              "title": "Sender Id"
            },
            "content": {
              "type": "string",
              "title": "Content"
            },
            "is_read": {
              "type": "boolean",
              "title": "Is Read"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "title": "Created At"
            }
          },
          "type": "object",
          "required": [
            "message_id",
            "sender_type",
            "sender_id",
            "content",
            "is_read",
            "created_at"
          ],
          "title": "Message"
        },
...
        "OrderPlaceFromCart": {
          "properties": {
            "delivery_date": {
              "type": "string",
              "title": "Delivery Date"
            },
            "delivery_window_start": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Delivery Window Start"
            },
            "delivery_window_end": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Delivery Window End"
            }
          },
          "type": "object",
          "required": [
            "delivery_date"
          ],
          "title": "OrderPlaceFromCart"
        },
...
        "OrderStatus": {
          "type": "string",
          "enum": [
            "placed",
            "under_risk",
            "waiting_for_customer_action",
            "picking",
            "delivering",
            "delivered",
            "cancelled"
          ],
          "title": "OrderStatus"
        },
...
        "OrderWithDetails": {
          "properties": {
            "delivery_date": {
              "type": "string",
              "title": "Delivery Date"
            },
            "delivery_window_start": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Delivery Window Start"
            },
            "delivery_window_end": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Delivery Window End"
            },
            "channel": {
              "type": "string",
              "title": "Channel",
              "default": "web"
            },
            "order_id": {
              "type": "string",
              "title": "Order Id"
            },
            "customer_id": {
              "type": "string",
              "title": "Customer Id"
            },
            "order_datetime": {
              "type": "string",
              "format": "date-time",
              "title": "Order Datetime"
            },
            "status": {
              "$ref": "#/components/schemas/OrderStatus"
            },
            "overall_order_risk": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Overall Order Risk"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "title": "Created At"
            },
            "updated_at": {
              "type": "string",
              "format": "date-time",
              "title": "Updated At"
            },
            "order_lines": {
              "items": {
                "$ref": "#/components/schemas/OrderLine"
              },
              "type": "array",
              "title": "Order Lines",
              "default": []
            },
            "order_substitutes": {
              "items": {
                "$ref": "#/components/schemas/OrderSubstitute"
              },
              "type": "array",
              "title": "Order Substitutes",
              "default": []
            },
            "tracking_history": {
              "items": {
                "$ref": "#/components/schemas/OrderTracking"
              },
              "type": "array",
              "title": "Tracking History",
              "default": []
            },
            "messages": {
              "items": {
                "$ref": "#/components/schemas/Message"
              },
              "type": "array",
              "title": "Messages",
              "default": []
            }
          },
          "type": "object",
          "required": [
            "delivery_date",
            "order_id",
            "customer_id",
            "order_datetime",
            "status",
            "created_at",
            "updated_at"
          ],
          "title": "OrderWithDetails"
        },
...
      },
      "securitySchemes": {
        "OAuth2PasswordBearer": {
          "type": "oauth2",
          "flows": {
            "password": {
              "scopes": {
  
              },
              "tokenUrl": "auth/login"
            }
          }
        }
      }
    }
  }